+++ NEWS.md
@@ -1,6 +1,21 @@
-# naomi 2.5.4
+# naomi 2.5.6
 
 * Add _simulated_ viral load testing data to example datasets.
+
+# naomi 2.5.5
+
+* Add argument `na.rm=` to `output_package()` to allow calculation of quantiles if there are missing values in the simulation. Default is `na.rm = FALSE` and `na.rm = TRUE` is to be used for debugging purposes only. Cases where missing values occur will usually indicate very poor model fits and issues that need to be addressed.
+
+# naomi 2.5.4
+
+Updates to PEPFAR Data Pack outputs.
+
+* Add age groups 50-54, 55-59, 60-64, 65+.
+* Edit format of PEPFAR data pack CSV:
+  - Add column `dataelement_uid`.
+  - Add `=""<>""` around age group values.
+* Change file name save to `pepfar_datapack_indicators_2022.csv`.
+>>>>>>> master
 * Separate input data aggregation and plot prepartion scripts
 * Remove `time_step` and add `year`, `quarter` and `calendar_quarter` in input 
 time series function outputs